update write length

+++ odbc/src/api/connect_attr_tests.rs
@@ -74,7 +74,7 @@ mod unit {
         unsafe fn modify_string_attr(value_ptr: Pointer, out_length: usize) -> String {
             input_text_to_string_w(
                 value_ptr as *const _,
-                out_length / std::mem::size_of::<WideChar>(),
+                (out_length / std::mem::size_of::<WideChar>()) as i32,
             )
         }
